requirements.txt

Flask==3.0.3
Keras==2.10.0
Pillow==9.5.0
numpy==1.23.5
gunicorn==22.0.0
